+++ tests/unit/test_networking_auth.py
@@ -4,11 +4,7 @@ import pytest
 
 import pip._internal.network.auth
 from pip._internal.network.auth import MultiDomainBasicAuth
-from tests.unit.test_operations_prepare import (
-    MockConnection,
-    MockRequest,
-    MockResponse,
-)
+from tests.lib.requests_mocks import MockConnection, MockRequest, MockResponse
 
 
 @pytest.mark.parametrize(["input_url", "url", "username", "password"], [
